% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_return.R
\name{ca_return_deleteBydata}
\alias{ca_return_deleteBydata}
\title{收款单按月删除--财务口径DMS}
\usage{
ca_return_deleteBydata(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
收款单按月删除--财务口径DMS
}
\examples{
ca_return_deleteBydata()
}
