% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_actual.R
\name{commision_actual_delete}
\alias{commision_actual_delete}
\title{清空佣金数据数据}
\usage{
commision_actual_delete(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
清空佣金数据数据
}
\examples{
commision_actual_delete()
}
