% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_actual.R
\name{commision_actual_insert}
\alias{commision_actual_insert}
\title{按月插入佣金数据数据}
\usage{
commision_actual_insert(dmstoken, year, MONTH)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
按月插入佣金数据数据
}
\examples{
commision_actual_insert()
}
