% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_actual_month.R
\name{commision_actual_month_select}
\alias{commision_actual_month_select}
\title{查询佣金月度汇总表——财务口径}
\usage{
commision_actual_month_select(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
查询佣金月度汇总表——财务口径
}
\examples{
commision_actual_month_select()
}
