% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_actual_month.R
\name{commision_actual_month_selectBydata}
\alias{commision_actual_month_selectBydata}
\title{按月计算佣金月度汇总表财务口径}
\usage{
commision_actual_month_selectBydata(dmstoken, year, month)
}
\arguments{
\item{month}{}
}
\value{
无返回值
}
\description{
按月计算佣金月度汇总表财务口径
}
\examples{
commision_actual_month_selectBydata()
}
