% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_actual.R
\name{commision_actual_select}
\alias{commision_actual_select}
\title{查询佣金表}
\usage{
commision_actual_select(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
查询佣金表
}
\examples{
commision_actual_select()
}
