% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_actual.R
\name{commision_actual_selectBydata}
\alias{commision_actual_selectBydata}
\title{按月查询佣金表}
\usage{
commision_actual_selectBydata(dmstoken, year, MONTH)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
按月查询佣金表
}
\examples{
commision_actual_selectBydata()
}
