% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_management.R
\name{commision_management_deleteBydate}
\alias{commision_management_deleteBydate}
\title{按月删除佣金数据数据}
\usage{
commision_management_deleteBydate(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
按月删除佣金数据数据
}
\examples{
commision_management_deleteBydate()
}
