% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_management_month.R
\name{commision_management_month_delete}
\alias{commision_management_month_delete}
\title{清空佣金月度汇总表——财务口径}
\usage{
commision_management_month_delete(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
清空佣金月度汇总表——财务口径
}
\examples{
commision_management_month_delete()
}
