% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commision_management.R
\name{commision_management_select}
\alias{commision_management_select}
\title{查询佣金管理口径表}
\usage{
commision_management_select(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
查询佣金管理口径表
}
\examples{
commision_management_select()
}
