% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_exchangerate.R
\name{exchangerate_deleteBydata}
\alias{exchangerate_deleteBydata}
\title{数据中台汇率体系表按月删除}
\usage{
exchangerate_deleteBydata(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
数据中台汇率体系表按月删除
}
\examples{
exchangerate_deleteBydata()
}
