% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_exchangerate.R
\name{exchangerate_jherp_selectBydata}
\alias{exchangerate_jherp_selectBydata}
\title{嘉好汇率体表按月查询径}
\usage{
exchangerate_jherp_selectBydata(erptoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
嘉好汇率体表按月查询径
}
\examples{
exchangerate_jherp_selectBydata()
}
