% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliQty_mngr.R
\name{fanliQty_mngr_delete}
\alias{fanliQty_mngr_delete}
\title{清空返利数量数据数据_管理口径}
\usage{
fanliQty_mngr_delete(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
清空返利数量数据数据_管理口径
}
\examples{
fanliQty_mngr_delete()
}
