% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliQty_mngr.R
\name{fanliQty_mngr_select}
\alias{fanliQty_mngr_select}
\title{查询返利数量表_管理口径}
\usage{
fanliQty_mngr_select(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
查询返利数量表_管理口径
}
\examples{
fanliQty_mngr_select()
}
