% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliQty_mngr.R
\name{fanliQty_mngr_selectBydata}
\alias{fanliQty_mngr_selectBydata}
\title{按月查询计算返利数量表_管理口径}
\usage{
fanliQty_mngr_selectBydata(dmstoken, year, MONTH)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
按月查询计算返利数量表_管理口径
}
\examples{
fanliQty_mngr_selectBydata()
}
