% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{grossProfit_management}
\alias{grossProfit_management}
\title{集团毛利表-管理口径}
\usage{
grossProfit_management(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
集团毛利表-管理口径
}
\examples{
grossProfit_management()
}
