% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grossProfit_management.R
\name{grossProfit_management_delete}
\alias{grossProfit_management_delete}
\title{集团毛利分许更新--删除客户行业表}
\usage{
grossProfit_management_delete(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
集团毛利分许更新--删除客户行业表
}
\examples{
grossProfit_management_delete()
}
