% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grossProfit_management.R
\name{grossProfit_management_jherp_insert}
\alias{grossProfit_management_jherp_insert}
\title{集团毛利分许更新--插入最新客户行业}
\usage{
grossProfit_management_jherp_insert(erptoken, dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
集团毛利分许更新--插入最新客户行业
}
\examples{
grossProfit_management_jherp_insert()
}
