% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grossProfit_management.R
\name{grossProfit_management_update}
\alias{grossProfit_management_update}
\title{集团毛利分许更新--更新最新客户行业}
\usage{
grossProfit_management_update(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
集团毛利分许更新--更新最新客户行业
}
\examples{
grossProfit_management_update()
}
