% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_outStock.R
\name{outstock_deleteBydata}
\alias{outstock_deleteBydata}
\title{销售出库按月删除--财务口径DMS}
\usage{
outstock_deleteBydata(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
销售出库按月删除--财务口径DMS
}
\examples{
outstock_deleteBydata()
}
