% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_outStock.R
\name{outstock_jherp_selectBydata}
\alias{outstock_jherp_selectBydata}
\title{销售出库--财务口径}
\usage{
outstock_jherp_selectBydata(erptoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
销售出库--财务口径
}
\examples{
outstock_jherp_selectBydata()
}
