% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_outstock_mngr.R
\name{outstock_mngr_deleteBydata}
\alias{outstock_mngr_deleteBydata}
\title{销售出库按月删除--财务口径DMS}
\usage{
outstock_mngr_deleteBydata(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
销售出库按月删除--财务口径DMS
}
\examples{
outstock_mngr_deleteBydata()
}
