% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_outstock_mngr.R
\name{outstock_mngr_jherp_selectBydata}
\alias{outstock_mngr_jherp_selectBydata}
\title{销售出库--管理口径}
\usage{
outstock_mngr_jherp_selectBydata(erptoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
销售出库--管理口径
}
\examples{
outstock_mngr_jherp_selectBydata()
}
