% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{rebate_detailAmt_actual}
\alias{rebate_detailAmt_actual}
\title{返利金额明细表--财务口径}
\usage{
rebate_detailAmt_actual(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
返利金额明细表--财务口径
}
\examples{
rebate_detailAmt_actual()
}
