% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliAmt_month.R
\name{rebate_detailAmt_actual_Bymonth}
\alias{rebate_detailAmt_actual_Bymonth}
\title{按月计算返利金额月度汇总表}
\usage{
rebate_detailAmt_actual_Bymonth(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
按月计算返利金额月度汇总表
}
\examples{
rebate_detailAmt_actual_Bymonth()
}
