% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliAmt_year.R
\name{rebate_detailAmt_actual_Byyear}
\alias{rebate_detailAmt_actual_Byyear}
\title{按年计算返利金额年度汇总表}
\usage{
rebate_detailAmt_actual_Byyear(dmstoken, year)
}
\arguments{
\item{year}{}
}
\value{
无返回值
}
\description{
按年计算返利金额年度汇总表
}
\examples{
rebate_detailAmt_actual_Byyear()
}
