% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanLiAmount.R
\name{rebate_detailAmt_actual_delete}
\alias{rebate_detailAmt_actual_delete}
\title{清空返利金额数据数据}
\usage{
rebate_detailAmt_actual_delete(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
清空返利金额数据数据
}
\examples{
rebate_detailAmt_actual_delete()
}
