% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanLiAmount.R
\name{rebate_detailAmt_actual_insert}
\alias{rebate_detailAmt_actual_insert}
\title{按月插入返利金额数据数据}
\usage{
rebate_detailAmt_actual_insert(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
按月插入返利金额数据数据
}
\examples{
rebate_detailAmt_actual_insert()
}
