% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanLiAmount.R
\name{rebate_detailAmt_actual_select}
\alias{rebate_detailAmt_actual_select}
\title{查询返利金额表}
\usage{
rebate_detailAmt_actual_select(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
查询返利金额表
}
\examples{
rebate_detailAmt_actual_select()
}
