% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanLiAmount.R
\name{rebate_detailAmt_actual_selectBydata}
\alias{rebate_detailAmt_actual_selectBydata}
\title{查询返利金额表按日期}
\usage{
rebate_detailAmt_actual_selectBydata(dmstoken, year, MONTH)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
查询返利金额表按日期
}
\examples{
rebate_detailAmt_actual_selectBydata()
}
