% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliAmt_year.R
\name{rebate_detailAmt_actual_selectByyear}
\alias{rebate_detailAmt_actual_selectByyear}
\title{查询返利金额年度汇总表}
\usage{
rebate_detailAmt_actual_selectByyear(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
查询返利金额年度汇总表
}
\examples{
rebate_detailAmt_actual_selectByyear()
}
