% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliQty_year.R
\name{rebate_detailQty_actual_Byyear}
\alias{rebate_detailQty_actual_Byyear}
\title{按月计算返利数量月度汇总表}
\usage{
rebate_detailQty_actual_Byyear(dmstoken, year)
}
\arguments{
\item{year}{}
}
\value{
无返回值
}
\description{
按月计算返利数量月度汇总表
}
\examples{
rebate_detailQty_actual_Byyear()
}
