% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanLiQty.R
\name{rebate_detailQty_actual_deleteBydate}
\alias{rebate_detailQty_actual_deleteBydate}
\title{按月删除返利数量数据数据}
\usage{
rebate_detailQty_actual_deleteBydate(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
按月删除返利数量数据数据
}
\examples{
rebate_detailQty_actual_deleteBydate()
}
