% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliQty_year.R
\name{rebate_detailQty_actual_deleteByyear}
\alias{rebate_detailQty_actual_deleteByyear}
\title{清空返利数量数据数据}
\usage{
rebate_detailQty_actual_deleteByyear(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
清空返利数量数据数据
}
\examples{
rebate_detailQty_actual_deleteByyear()
}
