% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliQty_month_mngr.R
\name{rebate_detailQty_actual_mngr_Bymonth}
\alias{rebate_detailQty_actual_mngr_Bymonth}
\title{按月计算返利数量月度汇总表管理口径}
\usage{
rebate_detailQty_actual_mngr_Bymonth(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
按月计算返利数量月度汇总表管理口径
}
\examples{
rebate_detailQty_actual_mngr_Bymonth()
}
