% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliQty_month_mngr.R
\name{rebate_detailQty_actual_mngr_deleteBymonth}
\alias{rebate_detailQty_actual_mngr_deleteBymonth}
\title{清空返利数量数据数据管理口径}
\usage{
rebate_detailQty_actual_mngr_deleteBymonth(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
清空返利数量数据数据管理口径
}
\examples{
rebate_detailQty_actual_mngr_deleteBymonth()
}
