% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliQty_month_mngr.R
\name{rebate_detailQty_actual_mngr_selectBymonth}
\alias{rebate_detailQty_actual_mngr_selectBymonth}
\title{查询返利数量月度汇总表管理口径}
\usage{
rebate_detailQty_actual_mngr_selectBymonth(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
查询返利数量月度汇总表管理口径
}
\examples{
rebate_detailQty_actual_mngr_selectBymonth()
}
