% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanLiQty.R
\name{rebate_detailQty_actual_selectBydata}
\alias{rebate_detailQty_actual_selectBydata}
\title{按月查询返利数量表}
\usage{
rebate_detailQty_actual_selectBydata(dmstoken, year, MONTH)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
按月查询返利数量表
}
\examples{
rebate_detailQty_actual_selectBydata()
}
