% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fanliQty_month.R
\name{rebate_detailQty_actual_selectBymonth}
\alias{rebate_detailQty_actual_selectBymonth}
\title{查询返利数量月度汇总表}
\usage{
rebate_detailQty_actual_selectBymonth(dmstoken)
}
\arguments{
\item{dmstoken}{}
}
\value{
无返回值
}
\description{
查询返利数量月度汇总表
}
\examples{
rebate_detailQty_actual_selectBymonth()
}
