% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_exchangerate.R
\name{rebate_exchangerate_delete}
\alias{rebate_exchangerate_delete}
\title{清空数据中台汇率体系表}
\usage{
rebate_exchangerate_delete(token)
}
\value{
无返回值
}
\description{
清空数据中台汇率体系表
}
\examples{
rebate_exchangerate_delete()
}
