% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebate_exchangerate.R
\name{rebate_exchangerate_select}
\alias{rebate_exchangerate_select}
\title{数据中台查询汇率体系表}
\usage{
rebate_exchangerate_select(token)
}
\value{
无返回值
}
\description{
数据中台查询汇率体系表
}
\examples{
rebate_exchangerate_select()
}
