% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{rebate_management}
\alias{rebate_management}
\title{返利金额明细表--管理口径}
\usage{
rebate_management(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
返利金额明细表--管理口径
}
\examples{
rebate_management()
}
