% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_receive.R
\name{RECEIVEBILL_jherp_selectBydata}
\alias{RECEIVEBILL_jherp_selectBydata}
\title{收款}
\usage{
RECEIVEBILL_jherp_selectBydata(erptoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
收款
}
\examples{
RECEIVEBILL_jherp_selectBydata()
}
