% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ca_return.R
\name{REFUNDBILL_jherp_selectBydata}
\alias{REFUNDBILL_jherp_selectBydata}
\title{收款退款}
\usage{
REFUNDBILL_jherp_selectBydata(erptoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
收款退款
}
\examples{
REFUNDBILL_jherp_selectBydata()
}
