% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_returnStock.R
\name{returnscock_jherp_selectBydata}
\alias{returnscock_jherp_selectBydata}
\title{销售退货--财务口径}
\usage{
returnscock_jherp_selectBydata(erptoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
销售退货--财务口径
}
\examples{
returnscock_jherp_selectBydata()
}
