% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_returnStock.R
\name{returnStock_deleteBydata}
\alias{returnStock_deleteBydata}
\title{销售退货按月删除--财务口径DMS}
\usage{
returnStock_deleteBydata(dmstoken, year, MONTH)
}
\arguments{
\item{MONTH}{}
}
\value{
无返回值
}
\description{
销售退货按月删除--财务口径DMS
}
\examples{
returnStock_deleteBydata()
}
