% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_commisionAmount.R
\name{rule_commisionAmount_add}
\alias{rule_commisionAmount_add}
\title{数据中台插入佣金规则}
\usage{
rule_commisionAmount_add(
  token,
  Fbillno,
  FSaleOrgName,
  FSaleMan,
  Fcountry,
  FStartDate,
  FEndDate,
  FCustomerName,
  FContidtionName,
  FCustMtrlName,
  FCommisionBy,
  FCommisionStandard,
  Fcurrency,
  FFiscalFee
)
}
\arguments{
\item{FFiscalFee}{}
}
\description{
数据中台插入佣金规则
}
\examples{
rule_commisionAmount_add()
}
