% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_commisionAmount.R
\name{rule_commisionAmount_delete}
\alias{rule_commisionAmount_delete}
\title{Title 数据中台佣金规则表删除}
\usage{
rule_commisionAmount_delete(token, FBillNo)
}
\arguments{
\item{FBillNo}{}
}
\description{
Title 数据中台佣金规则表删除
}
\examples{
rule_commisionAmount_delete()
}
