% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_commisionAmount.R
\name{rule_commisionAmount_view}
\alias{rule_commisionAmount_view}
\title{查询佣金规则表}
\usage{
rule_commisionAmount_view(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
查询佣金规则表
}
\examples{
rule_commisionAmount_view()
}
