% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_fanLiAmount.R
\name{rule_fanLiAmount_add}
\alias{rule_fanLiAmount_add}
\title{添加返利规则表}
\usage{
rule_fanLiAmount_add(
  token,
  FBillNo,
  FCustomerName,
  FIndexType,
  FStartRealAmount,
  FEndRealAmount,
  FUnit,
  FRebateRate,
  FEffectDate,
  FExpireDate
)
}
\arguments{
\item{FExpireDate}{}
}
\value{
无返回值
}
\description{
添加返利规则表
}
\examples{
rule_fanLiAmount_add()
}
