% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_fanLiAmount.R
\name{rule_fanLiAmount_delete}
\alias{rule_fanLiAmount_delete}
\title{Title}
\usage{
rule_fanLiAmount_delete(token, FBillNo)
}
\arguments{
\item{FBillNo}{}
}
\description{
Title
}
\examples{
rule_fanLiAmount_delete()
}
