% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_fanLiQty.R
\name{rule_fanLiQty_add}
\alias{rule_fanLiQty_add}
\title{添加返利规则表}
\usage{
rule_fanLiQty_add(
  token,
  FBillNo,
  FCustomerName,
  FIndexType,
  FIndexCustomerName,
  FStartRealQty,
  FEndRealQty,
  FUnit,
  FRebateProduct,
  FRebateType,
  FRebateRate,
  FEffectDate,
  FExpireDate
)
}
\arguments{
\item{FRebateType}{}
}
\value{
无返回值
}
\description{
添加返利规则表
}
\examples{
rule_fanLiQty_add()

}
