% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_fanLiQty.R
\name{rule_fanLiQty_delete}
\alias{rule_fanLiQty_delete}
\title{Title}
\usage{
rule_fanLiQty_delete(token, FBillNo)
}
\arguments{
\item{FBillNo}{}
}
\value{

}
\description{
Title
}
\examples{
rule_fanLiQty_delete()
}
