% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_fanLiQty.R
\name{rule_fanLiQty_view}
\alias{rule_fanLiQty_view}
\title{查询返利规则表}
\usage{
rule_fanLiQty_view(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
查询返利规则表
}
\examples{
rule_fanLiQty_view()
}
