% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_payment_month.R
\name{rule_payment_month_delete}
\alias{rule_payment_month_delete}
\title{Title 数据中台支付金额删除}
\usage{
rule_payment_month_delete(token, FContractNumber)
}
\arguments{
\item{token}{}
}
\value{

}
\description{
Title 数据中台支付金额删除
}
\examples{
rule_payment_month_delete()
}
