% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_payment_month.R
\name{rule_payment_month_view}
\alias{rule_payment_month_view}
\title{查询USD金额表}
\usage{
rule_payment_month_view(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
查询USD金额表
}
\examples{
rule_payment_month_view()
}
