% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_rebateFobamount.R
\name{rule_rebateFobamount_add}
\alias{rule_rebateFobamount_add}
\title{Title 增加计提依据}
\usage{
rule_rebateFobamount_add(token, FBillNo, FCountStandard)
}
\arguments{
\item{FCountStandard}{}
}
\description{
Title 增加计提依据
}
\examples{
rule_rebateFobamount_add()
}
