% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_rebateFobamount.R
\name{rule_rebateFobamount_delete}
\alias{rule_rebateFobamount_delete}
\title{Title 数据中台计提依据删除}
\usage{
rule_rebateFobamount_delete(token, FBillNo)
}
\arguments{
\item{FBillNo}{}
}
\description{
Title 数据中台计提依据删除
}
\examples{
rule_rebateFobamount_delete()
}
