% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule_rebateFobamount.R
\name{rule_rebateFobamount_view}
\alias{rule_rebateFobamount_view}
\title{查询FOB金额表}
\usage{
rule_rebateFobamount_view(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
查询FOB金额表
}
\examples{
rule_rebateFobamount_view()
}
