% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sal_outstock_mngr.R
\name{sal_outstock_mngr_select}
\alias{sal_outstock_mngr_select}
\title{查询销售出库管理口径}
\usage{
sal_outstock_mngr_select(token)
}
\value{
无返回值
}
\description{
查询销售出库管理口径
}
\examples{
sal_outstock_mngr_select()
}
