% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_management.R
\name{standard_management_delete}
\alias{standard_management_delete}
\title{标准成本分许更新--删除客户行业表}
\usage{
standard_management_delete(dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
标准成本分许更新--删除客户行业表
}
\examples{
standard_management_delete()
}
