% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_management.R
\name{standard_management_jherp_insert}
\alias{standard_management_jherp_insert}
\title{标准成本分许更新--插入最新客户行业}
\usage{
standard_management_jherp_insert(erptoken, dms_token)
}
\arguments{
\item{dms_token}{}
}
\value{
无返回值
}
\description{
标准成本分许更新--插入最新客户行业
}
\examples{
standard_management_jherp_insert()
}
