% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{standardCost_management}
\alias{standardCost_management}
\title{标准成本明细表}
\usage{
standardCost_management(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
标准成本明细表
}
\examples{
standardCost_management()
}
