% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{transportFee_actual}
\alias{transportFee_actual}
\title{运费明细表-财务口径}
\usage{
transportFee_actual(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
运费明细表-财务口径
}
\examples{
transportFee_actual()
}
