% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{transportFee_management}
\alias{transportFee_management}
\title{运费明细表-管理口径}
\usage{
transportFee_management(token)
}
\arguments{
\item{token}{}
}
\value{
无返回值
}
\description{
运费明细表-管理口径
}
\examples{
transportFee_management()
}
